Updating navigateToFrontend Command
There is an issue with the smoke build currently where the preview in the design section is not showing initially. Reloading the page seems to solve this problem.

I have included an interception for the preview request. If it returns a 500 then we will reload the page and continue with the test.

This solution should resolve the failing tests while investigating of the initial failing request occurs